requirements.txt
Flask
requests
python-nmap